Refactor garden search feature in templates and views
The garden search feature has been optimized through code refactoring in `gardens.html` and `views.py`. Instead of using the POST method, the search feature now uses the GET method for better performance and speed. An ID, `search-garden-form`, was introduced, and the JavaScript function `searchGardenSubmit()` is called upon the form submission. In `views.py`, the search was altered to search gardens based on the `garden-search` input rather than `card-search` and to return all gardens when no keyword is provided in the search box.