Add GardenOwner model and rename Address latitude field
Added a new GardenOwner model in 'gardens/migrations/' which includes gender, first name, and last name fields. Renamed the 'lattitude' field to 'latitude' in the address model found in 'gardens/models.py' and updated its references in the admin settings. Also, latitude and longitude values are automatically fetched and saved when an address is created or modified.