Reorganize URLs and add 'activities' path
The commit moves app and arrange_urlpatterns method to app/utils.py, and reorganizes the path to the gardens app in app/app/urls.py. Unused path 'activities' was removed from app/app/urls.py and a new 'activities' path was added, initiating the process of structuring the new 'activities' app. This cleanup and reorganising improves the maintainability of the URL patterns.